+++ src/Command/ICommand.ts
@@ -1,9 +1,12 @@
 import { OptionDefinition as ArgsOptionDefinition, CommandLineOptions } from "command-line-args";
 import { OptionDefinition as UsageOptionDefinition } from "command-line-usage";
 
+export type ICommandOption = ArgsOptionDefinition & UsageOptionDefinition;
+
 export default interface ICommand {
 	name: string;
-	optionList: (ArgsOptionDefinition & UsageOptionDefinition)[];
+	description: string;
+	optionList: ICommandOption[];
 	commands: ICommand[];
 	run(options: CommandLineOptions): Promise<void>;
 }